requirements.txt
transformers
open_clip_torch
torch
torchvision
pillow~=10.2.0
pinecone-client
requests~=2.31.0
lxml~=5.1.0

tqdm~=4.66.1
cog